Refactor API initialization and add CORS middleware
